refs #391, allow to send aircraft parts from GUI * GUI component for aircraft parts * remote aircraft selector component * Adjusted GUI for internals component * Enable / disable debug messages from GUI * Allow to init engines directly * Removed unused async sort in sequence
In same step fixed found issues in interpolator
* allow to set max rendered aircraft

+++ src/blackmisc/timestampobjectlist.cpp
@@ -119,7 +119,7 @@ namespace BlackMisc
     }
 
     template <class OBJ, class CONTAINER>
-    void ITimestampObjectList<OBJ, CONTAINER>::removeOlderThanNowMinusOffset(int offsetMs)
+    void ITimestampObjectList<OBJ, CONTAINER>::removeOlderThanNowMinusOffset(qint64 offsetMs)
     {
         const qint64 epoch = QDateTime::currentMSecsSinceEpoch() - offsetMs;
         this->container().removeIf([&](const OBJ & obj)
